Cambiados datos viejos tipo int a long

+++ src/main/java/com/sistema/application/repositories/IClienteRepository.java
@@ -13,7 +13,7 @@ import com.sistema.application.entities.Cliente;
 @Repository("clienteRepository")
 public interface IClienteRepository extends JpaRepository<Cliente, Serializable> {
 	//public abstract Cliente findById(long id);
-	public abstract Cliente findByNroCliente(int nroCliente);
+	public abstract Cliente findByNroCliente(long nroCliente);
 	public abstract List<Cliente> findByNombreAndApellido(String nombre, String apellido);
 	public abstract Cliente findByEmail(String email);
 }
