+++ src/main/java/com/sistema/application/services/ILocalService.java
@@ -11,6 +11,6 @@ public interface ILocalService {
 	
 	public LocalModel insertOrUpdate(LocalModel localModel);
 	
-	public boolean remove(int id);
+	public boolean remove(long id);
 	
 }
